Clean up code, add error paths
TestingUserDao is gone, DbUserDao is cleaned up, and now there are
error paths for everything using callbacks (lambdas are amazing).

We used callbacks instead of synchronous operations because it turns
out database operations are async, who knew?

Also because of this there's just an initialize method which throws
IOException and the rest don't anymore, instead they call an error
callback.